progressive/stage3/traffic-gen/pyproject.toml
[project]
name = "traffic-gen-stage3"
version = "1.0.0"
description = "Traffic generation script for Stage 3"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]
